added instructions: how to find "Object" class
Before, the only way to figure out that the "Objects" package is in "Kernal" is to look in the picture and see that, wherever "Objects" is, it is in a package a little above "Keymapping+Pragmas". I had to guess "Hey, 'Kernal' looks promising".
Now, there are step-by-step instructions on how to find everything.