Fix LaTeX compile of listings in Collections
The listings environment was changed to display a line in the margin.
Since the environment has an optional parameter, it tries to parse its
contents. If the code starts with either of `#{[`, this interferes and
breaks the LaTeX build.

The workaround is to set the language property, so that the LaTeX
environment finds a meaningful optional argument. Maybe making the
argument mandatory in the Pillar exporter would be a better solution.

That was not happening before with the original definition using
`listings.sty` directly, while the new one is through a wrapper. I think
'listings.sty' only looks for the optional argument immediately on the
`\begin{listing}` line, and not below, and the wrapper's parser is less
cautious.